+++ examples/video/liveness.cpp
@@ -1,8 +1,8 @@
 // An example of face services based on OpenCV camera streams.
 //
-// Author: Christian Kauten (ckauten@sensoryinc.com)
-//
 // Copyright (c) 2021 Sensory, Inc.
+//
+// Author: Christian Kauten (ckauten@sensoryinc.com)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -71,7 +71,7 @@ int main(int argc, const char** argv) {
         .help("DEVICE The ID of the OpenCV device to use.");
     parser.add_argument({ "-v", "--verbose" })
         .action("store_true")
-        .help("VERBOSE Produce verbose output during authentication.");
+        .help("VERBOSE Produce verbose output.");
     // Parse the arguments from the command line.
     const auto args = parser.parse_args();
     const auto HOSTNAME = args.get<std::string>("host");
@@ -125,36 +125,25 @@ int main(int argc, const char** argv) {
     sensory::service::OAuthService oauthService(config);
     sensory::token_manager::TokenManager<sensory::token_manager::InsecureCredentialStore> tokenManager(oauthService, keychain);
 
-    if (!tokenManager.hasToken()) {  // the device is not registered
-        // Generate a new clientID and clientSecret for this device
-        const auto credentials = tokenManager.hasSavedCredentials() ?
-            tokenManager.getSavedCredentials() : tokenManager.generateCredentials();
-
+    // Attempt to login and register the device if needed.
+    status = tokenManager.registerDevice([]() -> std::tuple<std::string, std::string> {
         std::cout << "Registering device with server..." << std::endl;
-
-        // Query the friendly device name
+        // Query the device name from the standard input.
         std::string name = "";
-        std::cout << "Device Name: ";
+        std::cout << "Device name: ";
         std::cin >> name;
-
-        // Query the shared pass-phrase
-        std::string password = "";
-        std::cout << "password: ";
-        std::cin >> password;
-
-        // Register this device with the remote host
-        sensory::api::v1::management::DeviceResponse registerResponse;
-        auto status = oauthService.registerDevice(&registerResponse,
-            name,
-            password,
-            credentials.id,
-            credentials.secret
-        );
-        if (!status.ok()) {  // the call failed, print a descriptive message
-            std::cout << "Failed to register device with\n\t" <<
-                status.error_code() << ": " << status.error_message() << std::endl;
-            return 1;
-        }
+        // Query the credential for the user from the standard input.
+        std::string credential = "";
+        std::cout << "Credential: ";
+        std::cin >> credential;
+        // Return the device name and credential as a tuple.
+        return {name, credential};
+    });
+    // Check the status code from the attempted registration.
+    if (!status.ok()) {  // the call failed, print a descriptive message
+        std::cout << "Failed to register device with\n\t" <<
+            status.error_code() << ": " << status.error_message() << std::endl;
+        return 1;
     }
 
     // ------ Create the video service -----------------------------------------
@@ -213,10 +202,8 @@ int main(int argc, const char** argv) {
     // (frame capture) and background (network stream processing) threads.
     std::mutex frameMutex;
 
-    // Create a thread to poll read requests in the background. Audio
-    // transcription has a bursty response pattern, so a locked read-write loop
-    // will not work with this service.
-    std::thread networkThread([&stream, &isLive, &alignmentCode, &frame, &frameMutex](){
+    // Create a thread to perform network IO in the background.
+    std::thread networkThread([&stream, &isLive, &alignmentCode, &frame, &frameMutex, &VERBOSE](){
         while (true) {
             std::vector<unsigned char> buffer;
             {  // Lock the mutex and encode the frame with JPEG into a buffer.
@@ -231,9 +218,11 @@ int main(int argc, const char** argv) {
             sensory::api::v1::video::LivenessRecognitionResponse response;
             if (!stream->Read(&response)) break;
             // Log information about the response to the terminal.
-            // std::cout << "Frame Response:" << std::endl;
-            // std::cout << "\tScore: "    << response.score() << std::endl;
-            // std::cout << "\tIs Alive: " << response.isalive() << std::endl;
+            if (VERBOSE) {
+                std::cout << "Frame Response:" << std::endl;
+                std::cout << "\tScore: "    << response.score() << std::endl;
+                std::cout << "\tIs Alive: " << response.isalive() << std::endl;
+            }
             // Set the authentication flag to the success of the response.
             isLive = response.isalive();
             alignmentCode = response.score() < 100 ?